// funderr-frontend/src/domain/types.ts
export interface CreditLine {
  id: string;
  codigo: string;
  nome: string;
  ativo: boolean;
  tetoFinanciamento: number;
  taxaJurosAnual: number;
  prazoMaxAnos: number;
  carenciaMaxAnos: number;
  percentualFinanciavelMax: number;
  percentualAterPadrao: number;
  observacoes: string;
  createdAt: string;
  updatedAt: string;
}